+++ hako-cli/src/cmd/serve.rs
@@ -17,6 +17,7 @@
 
 mod channel;
 mod client;
+mod deploy;
 mod proto;
 mod server;